refactor: move schema to validator
